requirements.txt
django-extensions
django
djangorestframework
twilio
djangorestframework-simplejwt